% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_horseshoe_fixed.R
\name{horseshoe.fixed}
\alias{horseshoe.fixed}
\title{Horseshoe prior model using a fixed shrinkage parameter}
\usage{
horseshoe.fixed(
  beta_MLE,
  Sigma_MLE,
  lambda2 = 1,
  a1 = 0.5,
  a2 = 0.5,
  a3 = 0.5,
  a4 = 0.5,
  b1 = 1,
  b2 = 1,
  Nsample = 500,
  store = 10,
  burnin = 100
)
}
\arguments{
\item{beta_MLE}{Initial MLE estimates.}

\item{Sigma_MLE}{Initial variance-covariance matrix.}

\item{lambda2}{Value of the shrinkage parameter, with default of 1.}

\item{a1}{Hyper-parameter of mixing distribution with the default value of 0.5.}

\item{a2}{Hyper-parameter of mixing distribution with the default value of 0.5.}

\item{a3}{Hyper-parameter of mixing distribution with the default value of 0.5.}

\item{a4}{Hyper-parameter of mixing distribution with the default value of 0.5.}

\item{b1}{Hyper-parameter of mixing distribution with the default value of 1.}

\item{b2}{Hyper-parameter of mixing distribution with the default value of 1.}

\item{Nsample}{Number of iterations of the Gibbs Sample with the default of 500.}

\item{store}{Fraction of the iteration to save from the sampler to improve quality of mixing, with default of 10.}

\item{burnin}{Number of first iterations to leave out, with default of 100.}
}
\value{
A list with a matrix of sampled coefficients, and a matrix of sampled hyper-parameter.
}
\description{
This function builds a Bayesian horseshoe model with fixed lambda parameter.
}
