% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_lasso_fixed.R
\name{lasso.fixed}
\alias{lasso.fixed}
\title{Lasso prior model using a fixed shrinkage parameter}
\usage{
lasso.fixed(
  beta_MLE,
  Sigma_MLE,
  Nsample = 500,
  store = 10,
  burnin = 100,
  lambda2 = 1
)
}
\arguments{
\item{beta_MLE}{Initial MLE estimates.}

\item{Sigma_MLE}{Initial variance-covariance matrix.}

\item{Nsample}{Number of iterations of the Gibbs Sample with the default of 500.}

\item{store}{Fraction of the iteration to save from the sampler to improve quality of mixing, with default of 10.}

\item{burnin}{Number of first iterations to leave out, with default of 100.}

\item{lambda2}{Value of the shrinkage parameter, with default of 1.}
}
\value{
A list with a matrix of sampled coefficients, and a matrix of sampled hyper-parameter.
}
\description{
This function builds a Bayesian lasso model with fixed lambda parameter.
}
