% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ridge_fixed.R
\name{ridge.fixed}
\alias{ridge.fixed}
\title{Ridge prior model using a fixed shrinkage parameter}
\usage{
ridge.fixed(beta_MLE, Sigma_MLE, lambda2 = 1, Nsample = 500)
}
\arguments{
\item{beta_MLE}{Initial MLE estimates.}

\item{lambda2}{Value of the shrinkage parameter, with default of 1.}

\item{Nsample}{Number of iterations of the Gibbs Sample with the default of 500.}
}
\value{
A matrix with Nsample rows and number of columns equal to the number of coefficients.
}
\description{
This function builds a Bayesian ridge model with fixed lambda parameter.
The posterior of the coefficients in this case is normal.
}
